refactor: move convertQueryParamsToUrlString to query.ts
