+++ libs/ts-rest/core/src/index.ts
@@ -6,3 +6,4 @@ export * from './lib/server-utils';
 export * from './lib/zod-utils';
 export * from './lib/status-codes';
 export * from './lib/paths';
+export * from './lib/query';
